feat: frontend custom hooks for coversion
